Fix: ignore rmtree() errors.
When deleting the temporary directory with shutil.rmtree(), a race condition
may cause the pidfile to be deleted after the files to delete are listed
but before they are actually deleted, causing an exception to be raised.
We now ignore these errors: we just want all the files to be deleted.